don't throw away etc/init.d/functions.sh as that causes misc misery
svn path=/main/branches/prefix/; revision=14409

+++ bin/misc-functions.sh
@@ -474,8 +474,8 @@ install_qa_check() {
 		# this code; many runscripts generate positives here, and we
 		# know they don't work (bug #196294) so as long as that one
 		# remains an issue, simply remove them as they won't work
-		# anyway
-		if [[ -d "${ED}"/etc/conf.d || -d "${ED}"/etc/init.d ]] ; then
+		# anyway, avoid etc/init.d/functions.sh from being thrown away
+		if [[ ( -d "${ED}"/etc/conf.d || -d "${ED}"/etc/init.d ) && ! -f "${ED}"/etc/init.d/functions.sh ]] ; then
 			ewarn "removed /etc/init.d and /etc/conf.d directories until bug #196294 has been resolved"
 			rm -Rf "${ED}"/etc/{conf,init}.d
 		fi
